fix string
