+++ ft_printf/configure_format_info.c
@@ -6,11 +6,12 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/03 14:17:21 by reasuke           #+#    #+#             */
-/*   Updated: 2024/02/23 22:45:24 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/02/24 12:36:23 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
+#include "ft_string.h"
 
 static void	_set_flags(t_format_info *fi, t_format_result *fr)
 {
