+++ ft_printf/print_by_format.c
@@ -6,11 +6,12 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/03 16:13:31 by reasuke           #+#    #+#             */
-/*   Updated: 2024/02/24 00:01:37 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/02/24 12:36:12 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
+#include "ft_string.h"
 
 static void	_print_string_by_format(t_format_info *fi,
 				t_format_result *fr, va_list *ap, int fd)
