+++ ft_string/ft_strchr.c
@@ -1,29 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strrchr.c                                       :+:      :+:    :+:   */
+/*   ft_strchr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/02 14:56:21 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/03 12:06:38 by reasuke          ###   ########.fr       */
+/*   Created: 2023/09/02 14:53:37 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:31:37 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_string.h"
 
-char	*ft_strrchr(const char *s, int c)
+char	*ft_strchr(const char *s, int c)
 {
-	const char	*ret;
-
-	ret = NULL;
 	while (*s)
 	{
 		if ((unsigned char)*s == (unsigned char)c)
-			ret = s;
+			return ((char *)s);
 		s++;
 	}
 	if ((unsigned char)*s == (unsigned char)c)
 		return ((char *)s);
-	return ((char *)ret);
+	return (NULL);
 }
