+++ ft_string/ft_strcmp.c
