+++ ft_string/ft_strdup.c
@@ -1,24 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_free_strs.c                                     :+:      :+:    :+:   */
+/*   ft_strdup.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/03 15:06:58 by reasuke           #+#    #+#             */
-/*   Updated: 2023/11/03 15:10:30 by reasuke          ###   ########.fr       */
+/*   Created: 2023/09/02 17:49:30 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:32:23 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_string.h"
 
-void	*ft_free_strs(char **strs)
+char	*ft_strdup(const char *s1)
 {
-	char	**tmp;
+	char	*ret;
+	size_t	s1_len;
 
-	tmp = strs;
-	while (*tmp)
-		free(*tmp++);
-	free(strs);
-	return (NULL);
+	s1_len = ft_strlen(s1);
+	ret = malloc(s1_len + 1);
+	if (!ret)
+		return (NULL);
+	ft_strlcpy(ret, s1, s1_len + 1);
+	return (ret);
 }
