+++ ft_string/ft_striteri.c
@@ -6,11 +6,11 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/04 01:27:36 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/05 17:39:32 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/02/24 12:32:28 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_string.h"
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
