+++ ft_string/ft_strjoin.c
@@ -1,35 +1,35 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strmapi.c                                       :+:      :+:    :+:   */
+/*   ft_strjoin.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/03 22:50:20 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/06 00:23:36 by reasuke          ###   ########.fr       */
+/*   Created: 2023/09/03 12:49:03 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:32:31 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_string.h"
 
-char	*ft_strmapi(const char *s, char (*f)(unsigned int, char))
+char	*ft_strjoin(const char *s1, const char *s2)
 {
 	char	*ret;
-	size_t	len;
-	size_t	i;
+	size_t	s1_len;
+	size_t	s2_len;
 
-	if (!s)
+	if (!s1 && !s2)
 		return (NULL);
-	len = ft_strlen(s);
-	ret = malloc(len + 1);
+	else if (!s1)
+		return (ft_strdup(s2));
+	else if (!s2)
+		return (ft_strdup(s1));
+	s1_len = ft_strlen(s1);
+	s2_len = ft_strlen(s2);
+	ret = malloc(s1_len + s2_len + 1);
 	if (!ret)
-		return (ret);
-	i = 0;
-	while (i < len)
-	{
-		ret[i] = f(i, s[i]);
-		i++;
-	}
-	ret[len] = '\0';
+		return (NULL);
+	ft_strlcpy(ret, s1, s1_len + 1);
+	ft_strlcpy(ret + s1_len, s2, s2_len + 1);
 	return (ret);
 }
