+++ ft_string/ft_strlen.c
@@ -1,27 +1,23 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strlcat.c                                       :+:      :+:    :+:   */
+/*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/02 17:37:54 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/28 20:10:50 by reasuke          ###   ########.fr       */
+/*   Created: 2023/09/02 15:17:21 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:32:40 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_string.h"
 
-size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
+size_t	ft_strlen(const char *s)
 {
-	size_t	dst_len;
+	size_t	len;
 
-	if (dstsize == 0)
-		return (ft_strlen(src));
-	dst_len = ft_strlen(dst);
-	if (dstsize <= dst_len)
-		return (ft_strlen(src) + dstsize);
-	while (*dst)
-		dst++;
-	return (ft_strlcpy(dst, src, dstsize - dst_len) + dst_len);
+	len = 0;
+	while (s[len])
+		len++;
+	return (len);
 }
