+++ ft_string/ft_strmapi.c
@@ -1,35 +1,35 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_substr.c                                        :+:      :+:    :+:   */
+/*   ft_strmapi.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/03 12:34:33 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/05 17:38:17 by reasuke          ###   ########.fr       */
+/*   Created: 2023/09/03 22:50:20 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:32:44 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_string.h"
 
-char	*ft_substr(const char *s, unsigned int start, size_t len)
+char	*ft_strmapi(const char *s, char (*f)(unsigned int, char))
 {
 	char	*ret;
-	size_t	s_len;
-	size_t	ret_len;
+	size_t	len;
+	size_t	i;
 
 	if (!s)
 		return (NULL);
-	s_len = ft_strlen(s);
-	if (s_len <= start)
-		return (ft_strdup(""));
-	if (len < s_len - start + 1)
-		ret_len = len;
-	else
-		ret_len = s_len - start + 1;
-	ret = malloc(ret_len + 1);
+	len = ft_strlen(s);
+	ret = malloc(len + 1);
 	if (!ret)
-		return (NULL);
-	ft_strlcpy(ret, s + start, ret_len + 1);
+		return (ret);
+	i = 0;
+	while (i < len)
+	{
+		ret[i] = f(i, s[i]);
+		i++;
+	}
+	ret[len] = '\0';
 	return (ret);
 }
