+++ ft_string/ft_substr.c
@@ -1,32 +1,35 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strnstr.c                                       :+:      :+:    :+:   */
+/*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/02 15:00:49 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/05 17:31:07 by reasuke          ###   ########.fr       */
+/*   Created: 2023/09/03 12:34:33 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:33:14 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_string.h"
 
-char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
+char	*ft_substr(const char *s, unsigned int start, size_t len)
 {
-	size_t	needle_len;
+	char	*ret;
+	size_t	s_len;
+	size_t	ret_len;
 
-	if (!*needle)
-		return ((char *)haystack);
-	if (len == 0)
+	if (!s)
 		return (NULL);
-	needle_len = ft_strlen(needle);
-	while (*haystack && needle_len <= len)
-	{
-		if (!ft_strncmp(haystack, needle, needle_len))
-			return ((char *)haystack);
-		haystack++;
-		len--;
-	}
-	return (NULL);
+	s_len = ft_strlen(s);
+	if (s_len <= start)
+		return (ft_strdup(""));
+	if (len < s_len - start + 1)
+		ret_len = len;
+	else
+		ret_len = s_len - start + 1;
+	ret = malloc(ret_len + 1);
+	if (!ret)
+		return (NULL);
+	ft_strlcpy(ret, s + start, ret_len + 1);
+	return (ret);
 }
