--- string/ft_strtrim.c
@@ -1,36 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_strtrim.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/03 23:01:26 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/23 02:52:31 by reasuke          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-char	*ft_strtrim(const char *s1, const char *set)
-{
-	char	*ret;
-	size_t	begin;
-	size_t	end;
-
-	if (!s1 || !set)
-		return (NULL);
-	begin = 0;
-	while (s1[begin] && ft_strchr(set, s1[begin]))
-		begin++;
-	if (begin == ft_strlen(s1))
-		return (ft_strdup(""));
-	end = ft_strlen(s1) - 1;
-	while (ft_strchr(set, s1[end]))
-		end--;
-	ret = malloc(end - begin + 2);
-	if (!ret)
-		return (NULL);
-	ft_strlcpy(ret, s1 + begin, end - begin + 2);
-	return (ret);
-}
